feat: add ConsoleLink creation page
Adds a form for creating `ConsoleLink` resources that supports
`ApplicationMenu`, `HelpMenu`, and `UserMenu` links. The page does not
yet support specifying an icon for `ApplicationMenu` links or creating
`NamespaceDashboard` links. This form will be used when creating a
`ConsoleLink` from the "Customization" page or anywhere else in console.

Introduces PatternFly components to the plugin and moves some k8s
utilities into a separate package.

I did need to workaround a display issue with PatternFly buttons where
the button CSS was included twice, breaking the console layout by
changing the positioning of the hidden skip to content button in the
masthead.

The plugin uses `consoleFetch` directly since `k8sCreate` is not yet
exposed as an API. Additionally, I hard-code some paths. Those should be
replaced by plugin SDK utilities when available.